Add logic to hide Header when visiting login route
A new state is created, which is assigned the value of the Header main
container display to compare it with the current value of the display.

When visiting the login route the display is set to 'none', and
if when visiting a different route it is still 'none', it is set to
'flex'.