Change if logic in useStates for login process
Previously, the code within the useState hooks was executed only if
the corresponding state was not null. This resulted in additional
indentation. To eliminate this, the conditions now check if the state is
null and return undefined if it is. As a result, the extra indentation
is no longer necessary.